backend/requirements.txt
Flask
requests
langchain
langchain_openai
python-dotenv
gunicorn
